[chore #160547960] rename utilities to validations and update imports
